Add username validation to server

+++ routes/ideas.js
@@ -44,33 +44,50 @@ router.post('/', async (req, res) => {
 // Update an idea
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-            {
-                $set: {
-                    text: req.body.text,
-                    tag: req.body.tag,
+        const idea = await Idea.findById(req.params.id);
+        
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                {
+                    $set: {
+                        text: req.body.text,
+                        tag: req.body.tag,
+                    }
+                },
+                {
+                    new: true,
                 }
-            },
-            {
-                new: true,
-            }
-        );
-        res.json({ success: true, data: updatedIdea });
+            );
+            res.json({ success: true, data: updatedIdea });
+        } else {
+            // Usernames don't match
+            res.status(403).json({ success: false, error: 'You are not authorised to update this resource'});
+        }
+        
     } catch (error) {
         console.log(error);
-        res.status(404).json({ success: false, error: 'Resource not found'});
+        res.status(500).json({ success: false, error: 'Something went wrong'});
     }
 });
 
 // Delete an idea
 router.delete('/:id', async (req, res) => {
     try {
-        const idea = await Idea.findByIdAndDelete(req.params.id);
-        res.json({ success: true, data: idea });
+        const idea = await Idea.findById(req.params.id);
+        
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            res.json({ success: true, data: {} });
+        } else {
+            // Usernames don't match
+            res.status(403).json({ success: false, error: 'You are not authorised to delete this resource'});
+        }
     } catch (error) {
         console.log(error);
-        res.status(404).json({ success: false, error: 'Resource not found'});
+        res.status(500).json({ success: false, error: 'Something went wrong'});
     }
 });
